perf: GetInterfaceTrait is added to implement the interface get method
