+++ src/LionDatabase/Drivers/PostgreSQL.php
@@ -5,11 +5,12 @@ declare(strict_types=1);
 namespace Lion\Database\Drivers;
 
 use Lion\Database\Connection;
+use Lion\Database\Driver;
 use Lion\Database\Helpers\ConnectionInterfaceTrait;
+use Lion\Database\Helpers\GetInterfaceTrait;
 use Lion\Database\Helpers\QueryInterfaceTrait;
 use Lion\Database\Helpers\RunInterfaceTrait;
 use Lion\Database\Helpers\TransactionInterfaceTrait;
-use Lion\Database\Interface\DatabaseCapsuleInterface;
 use Lion\Database\Interface\DatabaseConfigInterface;
 use Lion\Database\Interface\QueryInterface;
 use Lion\Database\Interface\ReadDatabaseDataInterface;
@@ -31,6 +32,9 @@ use stdClass;
  * * Optimization for PostgreSQL: Designed specifically to work with PostgreSQL,
  *   guaranteeing compatibility and optimization with this DBMS.
  *
+ * @property string $databaseMethod [Defines the database connection method to
+ * use]
+ *
  * @package Lion\Database\Drivers
  */
 class PostgreSQL extends Connection implements
@@ -41,74 +45,22 @@ class PostgreSQL extends Connection implements
     TransactionInterface
 {
     use ConnectionInterfaceTrait;
+    use GetInterfaceTrait;
     use QueryInterfaceTrait;
     use RunInterfaceTrait;
     use TransactionInterfaceTrait;
 
     /**
-     * {@inheritdoc}
+     * Defines the database connection method to use
+     *
+     * This property determines which connection method to use in the `trait` to
+     * perform database operations. Allowed values are `mysql` or `postgresql`,
+     * depending on the database being used. The class using the `trait` must
+     * set this value to define the connection type
+     *
+     * @var string $databaseMethod
      */
-    public static function get(): stdClass|array|DatabaseCapsuleInterface
-    {
-        return parent::postgresql(function (): stdClass|array|DatabaseCapsuleInterface {
-            $responses = [];
-
-            self::$listSql = array_map(
-                fn ($value) => trim($value),
-                array_filter(explode(';', trim(self::$sql)), fn ($value) => trim($value) != '')
-            );
-
-            $codes = array_keys(self::$fetchMode);
-
-            foreach (self::$listSql as $key => $sql) {
-                self::prepare($sql);
-
-                $code = $codes[$key] ?? null;
-
-                if ($code != null && isset(self::$dataInfo[$code])) {
-                    self::bindValue($code);
-                }
-
-                if ($code != null && isset(self::$fetchMode[$code])) {
-                    $getFetch = self::$fetchMode[$codes[$key]];
-
-                    if (is_array($getFetch)) {
-                        self::$stmt->setFetchMode($getFetch[0], $getFetch[1]);
-                    } else {
-                        self::$stmt->setFetchMode(self::$fetchMode[$codes[$key]]);
-                    }
-                }
-
-                self::$stmt->execute();
-
-                $request = self::$stmt->fetch();
-
-                if (!$request) {
-                    if (count(self::$fetchMode) > 1) {
-                        $responses[] = (object) [
-                            'code' => 200,
-                            'status' => 'success',
-                            'message' => 'no data available',
-                        ];
-                    } else {
-                        $responses = (object) [
-                            'code' => 200,
-                            'status' => 'success',
-                            'message' => 'no data available',
-                        ];
-                    }
-                } else {
-                    if (count(self::$fetchMode) > 1) {
-                        $responses[] = $request;
-                    } else {
-                        $responses = $request;
-                    }
-                }
-            }
-
-            return $responses;
-        });
-    }
+    private static string $databaseMethod = Driver::POSTGRESQL;
 
     /**
      * {@inheritdoc}
